# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Locação API",
    "version": "1.0.0",
    "description": "API para gerenciamento de locações e brinquedos."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor Local"
    }
  ],
  "paths": {
...
    "/clientes": {
      "get": {
        "summary": "Listar todos os clientes",
        "operationId": "getClientes",
        "responses": {
          "200": {
            "description": "Lista de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteResponseDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo cliente",
        "operationId": "createCliente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClienteDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/clientes/{cpf}": {
      "parameters": [
        {
          "name": "cpf",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[0-9]{11}$"
          }
        }
      ],
      "get": {
        "summary": "Obter um cliente pelo CPF",
        "operationId": "getClienteByCpf",
        "responses": {
          "200": {
            "description": "Cliente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar um cliente",
        "operationId": "updateCliente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClienteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/funcionarios": {
      "get": {
        "summary": "Listar todos os funcionários",
        "operationId": "getFuncionarios",
        "responses": {
          "200": {
            "description": "Lista de funcionários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FuncionarioResponseDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo funcionário",
        "operationId": "createFuncionario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFuncionarioDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Funcionário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/funcionarios/{cpf}": {
      "parameters": [
        {
          "name": "cpf",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[0-9]{11}$"
          }
        }
      ],
      "get": {
        "summary": "Obter um funcionário pelo CPF",
        "operationId": "getFuncionarioByCpf",
        "responses": {
          "200": {
            "description": "Funcionário encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar um funcionário",
        "operationId": "updateFuncionario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFuncionarioDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Funcionário atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuncionarioResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Excluir um funcionário",
        "operationId": "deleteFuncionario",
        "responses": {
          "204": {
            "description": "Funcionário removido"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateClienteDTO": {
        "type": "object",
        "required": ["cpf", "nome", "endereco", "data_nascimento", "telefone"],
        "properties": {
          "cpf": {
            "type": "string",
            "pattern": "^[0-9]{11}$",
            "example": "12345678901"
          },
          "nome": {
            "type": "string",
            "example": "João Silva"
          },
          "endereco": {
            "type": "string",
            "example": "Rua Exemplo, 123"
          },
          "data_nascimento": {
            "type": "string",
            "format": "date",
            "example": "1990-01-01"
          },
          "telefone": {
            "type": "string",
            "example": "11987654321"
          }
        },
        "additionalProperties": false
      },
      "UpdateClienteDTO": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "endereco": {
            "type": "string"
          },
          "data_nascimento": {
            "type": "string",
            "format": "date"
          },
          "telefone": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ClienteResponseDTO": {
        "type": "object",
        "properties": {
          "cpf": {
            "type": "string",
            "pattern": "^[0-9]{11}$"
          },
          "nome": {
            "type": "string"
          },
          "endereco": {
            "type": "string"
          },
          "data_nascimento": {
            "type": "string",
            "format": "date"
          },
          "telefone": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
      "CreateFuncionarioDTO": {
        "type": "object",
        "required": ["cpf", "nome", "telefone", "funcao", "senha"],
        "properties": {
          "cpf": {
            "type": "string",
            "pattern": "^[0-9]{11}$",
            "example": "12345678901"
          },
          "nome": {
            "type": "string",
            "example": "Maria Souza"
          },
          "telefone": {
            "type": "string",
            "example": "11987654321"
          },
          "funcao": {
            "type": "string",
            "enum": [
              "GERENTE",
              "CAIXA",
              "ALMOXARIFE",
              "AGENTE_LOCACAO",
              "ANALISTA_CADASTRO"
            ],
            "example": "GERENTE"
          },
          "senha": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "example": "senhaSegura123!"
          }
        },
        "additionalProperties": false
      },
      "UpdateFuncionarioDTO": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "funcao": {
            "type": "string",
            "enum": [
              "GERENTE",
              "CAIXA",
              "ALMOXARIFE",
              "AGENTE_LOCACAO",
              "ANALISTA_CADASTRO"
            ]
          },
          "senha": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "FuncionarioResponseDTO": {
        "type": "object",
        "properties": {
          "cpf": {
            "type": "string",
            "pattern": "^[0-9]{11}$"
          },
          "nome": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "funcao": {
            "type": "string",
            "enum": [
              "GERENTE",
              "CAIXA",
              "ALMOXARIFE",
              "AGENTE_LOCACAO",
              "ANALISTA_CADASTRO"
            ]
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}